*migration fix

+++ app/database/migrations/2015_04_04_094611_create_service_account_type_table.php
@@ -17,7 +17,7 @@ class CreateServiceAccountTypeTable extends Migration {
 
             $table->bigIncrements('id')->unsigned();
             $table->string('name', 255);
-            $table->smallInteger('limit_application')->unsigned();
+            $table->bigInteger('limit_application')->unsigned();
             $table->bigInteger('limit_keyring')->unsigned();
             $table->bigInteger('limit_pki')->unsigned();
             $table->bigInteger('limit_auth')->unsigned();
